[refactor] Update `GetMarketChartUseCase` and `GetMarketDetailUseCase` to use new Resource type
