+++ domain/market/src/main/java/ir/composenews/domain/use_case/GetMarketChartUseCase.kt
@@ -3,13 +3,15 @@
 package ir.composenews.domain.use_case
 
 import ir.composenews.domain.model.Chart
-import ir.composenews.domain.model.Resource
 import ir.composenews.domain.repository.MarketRepository
+import ir.composenews.ktor.Errors
+import ir.composenews.ktor.Resource
 import kotlinx.coroutines.flow.Flow
 import javax.inject.Inject
 
 open class GetMarketChartUseCase @Inject constructor(
     private val repository: MarketRepository,
 ) {
-    open operator fun invoke(id: String): Flow<Resource<Chart>> = repository.fetchChart(id = id)
+    open operator fun invoke(id: String): Flow<Resource<Chart, Errors>> =
+        repository.fetchChart(id = id)
 }
